feat(extractor): add XHamster extractor and refactor shared utilities
- Add rdlp-crypto crate with PRNG-based URL decryption for XHamster
- Add XHamster extractor with encrypted URL support
- Split HLS downloader into modules (merge, playlist, segment, types)
- Add shared utilities to base/common.rs:
  - HTML_TAG_PATTERN and clean_html_tags()
  - parse_duration() for ISO 8601, colon-separated, and text formats
  - parse_text_duration() for "28min 18sec" format
- Add base/kvs.rs for KVS flashvars parsing (used by XTits and future KVS sites)
- Refactor extractors to use shared utilities:
  - pornhub/utils.rs uses clean_html_tags()
  - xhamster/utils.rs uses clean_html_tags() and parse_duration()
  - xtits uses KvsFlashvars and parse_text_duration()
- Extract HTTP downloader constants to config.rs
- Add with_retry helper to reduce retry boilerplate
- Merge merge_chunks/append_chunks into merge_chunks_with_mode
