+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "crates/rdlp-core",
     "crates/rdlp-http",
     "crates/rdlp-security",
+    "crates/rdlp-crypto",
     "crates/rdlp-extractor",
     "crates/rdlp-downloader",
     "crates/rdlp-jsinterp",
@@ -85,6 +86,7 @@ once_cell = "1.20"
 futures = "0.3"
 futures-util = "0.3"
 bytes = "1.9"
+hex = "0.4"
 
 # Testing
 mockito = "1.5"
