+++ crates/rdlp-crypto/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "rdlp-crypto"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+description = "Cryptographic utilities for rdlp including PRNG-based URL decryption"
+
+[dependencies]
+log = { workspace = true }
+once_cell = { workspace = true }
+regex = { workspace = true }
+url = { workspace = true }
+hex = { workspace = true }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["test-util"] }
+
+[lints]
+workspace = true
