--- crates/rdlp-downloader/src/hls.rs
@@ -1,1059 +1,0 @@
-use async_trait::async_trait;
-use backon::Retryable;
-use futures::stream::{self, StreamExt};
-use log::{debug, info, warn};
-use rdlp_core::{
-    DownloadStats, Downloader, ProgressCallback, RdlpError, Result, RetryConfig,
-    is_retryable_error,
-};
-use std::collections::{HashMap, HashSet};
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::time::{Duration, Instant};
-use tokio::fs::File;
-use tokio::io::{AsyncWriteExt, BufWriter};
-use tokio::sync::Mutex;
-use tracing::instrument;
-
-use crate::hls_state::HlsDownloadState;
-use crate::http::HttpDownloader;
-use crate::progress::{ProgressMetrics, ProgressReporterConfig, spawn_progress_reporter};
-
-/// EXT-X-MAP initialization segment info (fMP4 streams)
-///
-/// Per the HLS spec, `EXT-X-MAP` applies to every segment after it until
-/// the next `EXT-X-MAP` tag. A playlist may therefore contain multiple
-/// init segments (e.g. codec change mid-stream, ad insertion).
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-struct InitSegmentInfo {
-    /// Fully-resolved URL of the initialization segment
-    url: String,
-    /// Optional byte range `(length, offset)` — when the init data is
-    /// packed inside a larger resource (e.g. the same URI as the segments)
-    byte_range: Option<(u64, Option<u64>)>,
-}
-
-/// Information about an HLS segment including its duration
-#[derive(Clone, Debug)]
-struct SegmentInfo {
-    /// Segment URL
-    url: String,
-    /// Segment duration in seconds (from EXTINF)
-    duration: f64,
-    /// The EXT-X-MAP init segment that applies to this segment (if any)
-    init_segment: Option<InitSegmentInfo>,
-}
-
-/// Result of parsing an HLS playlist
-#[derive(Clone, Debug)]
-struct PlaylistParseResult {
-    /// Media segments (each carrying its own init segment reference)
-    segments: Vec<SegmentInfo>,
-}
-
-/// HLS (HTTP Live Streaming) downloader
-///
-/// Downloads HLS streams by parsing m3u8 playlists and downloading segments
-/// in parallel using the HTTP downloader infrastructure.
-///
-/// # Architecture
-///
-/// 1. Parse m3u8 playlist to extract segment URLs
-/// 2. Download segments in parallel using `buffer_unordered`
-/// 3. Merge segments into final video file
-/// 4. Clean up temporary segment files
-///
-/// # Performance
-///
-/// - Default: 8 concurrent segment downloads
-/// - Typical: 500 MB video in 60-90 seconds
-/// - Bottleneck: Server throttling (not client)
-///
-/// # Example
-///
-/// ```rust,no_run
-/// use rdlp_downloader::HlsDownloader;
-/// use rdlp_core::Downloader;
-/// use std::path::Path;
-///
-/// # async fn example() -> rdlp_core::Result<()> {
-/// let downloader = HlsDownloader::new();
-/// downloader.download_to_file(
-///     "https://example.com/playlist.m3u8",
-///     Path::new("video.mp4"),
-///     None
-/// ).await?;
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Clone)]
-pub struct HlsDownloader {
-    http_downloader: HttpDownloader,
-    concurrent_segments: usize,
-    buffer_size: usize,
-    retry_config: Arc<RetryConfig>,
-    /// Expected total size for progress reporting (set externally)
-    expected_size: Option<u64>,
-    /// Total download timeout (entire operation must complete within this)
-    download_timeout: Duration,
-    /// Merge operation timeout (segment merge must complete within this)
-    merge_timeout: Duration,
-    /// Maximum number of segment failures before aborting the download
-    max_segment_failures: usize,
-}
-
-impl HlsDownloader {
-    /// Create a new HLS downloader with default settings
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            http_downloader: HttpDownloader::new(),
-            concurrent_segments: 8,       // Default: 8 parallel segments
-            buffer_size: 2 * 1024 * 1024, // 2 MB buffer for merging
-            retry_config: Arc::new(RetryConfig::default_config()),
-            expected_size: None,
-            download_timeout: Duration::from_secs(3600), // 1 hour
-            merge_timeout: Duration::from_secs(1800),    // 30 min
-            max_segment_failures: 3,
-        }
-    }
-
-    /// Set the HTTP downloader to use
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_http_downloader(mut self, http: HttpDownloader) -> Self {
-        self.http_downloader = http;
-        self
-    }
-
-    /// Set number of concurrent segment downloads
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_concurrent_segments(mut self, count: usize) -> Self {
-        self.concurrent_segments = count.max(1);
-        self
-    }
-
-    /// Set buffer size for segment merging
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_buffer_size(mut self, size: usize) -> Self {
-        self.buffer_size = size;
-        self
-    }
-
-    /// Set retry configuration
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_retry_config(mut self, config: RetryConfig) -> Self {
-        self.retry_config = Arc::new(config);
-        self
-    }
-
-    /// Set expected total size for progress reporting
-    ///
-    /// This allows the progress bar to show accurate percentage and ETA
-    /// even though HLS streams don't have a known total size upfront.
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_expected_size(mut self, size: u64) -> Self {
-        self.expected_size = Some(size);
-        self
-    }
-
-    /// Set total download timeout
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_download_timeout(mut self, timeout: Duration) -> Self {
-        self.download_timeout = timeout;
-        self
-    }
-
-    /// Set merge operation timeout
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_merge_timeout(mut self, timeout: Duration) -> Self {
-        self.merge_timeout = timeout;
-        self
-    }
-
-    /// Set maximum number of segment failures before aborting
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_max_segment_failures(mut self, max: usize) -> Self {
-        self.max_segment_failures = max;
-        self
-    }
-
-    /// Set extra HTTP headers sent with every request (delegates to inner HttpDownloader)
-    #[must_use = "builder methods consume self and return a new instance"]
-    pub fn with_extra_headers(mut self, headers: Option<&std::collections::HashMap<String, String>>) -> Self {
-        self.http_downloader = self.http_downloader.with_extra_headers(headers);
-        self
-    }
-
-    /// Download a single HLS segment with retry logic using backon
-    ///
-    /// Handles network errors, timeouts, and expired URLs by retrying with exponential backoff.
-    ///
-    /// # Arguments
-    /// * `idx` - Segment index (for logging)
-    /// * `url` - Segment URL
-    /// * `segment_path` - Path to save segment
-    /// * `progress` - Shared progress counter
-    ///
-    /// # Returns
-    /// * `Ok((index, path, bytes))` - Successfully downloaded segment
-    /// * `Err(_)` - Failed after all retries
-    #[instrument(skip(self, progress), fields(segment = idx))]
-    async fn download_segment_with_retry(
-        &self,
-        idx: usize,
-        url: String,
-        segment_path: PathBuf,
-        progress: Arc<AtomicU64>,
-    ) -> Result<(usize, PathBuf, u64)> {
-        let http_client = self.http_downloader.client().clone();
-        let rate_limiter = self.http_downloader.rate_limiter.clone();
-        let buffer_size = self.buffer_size;
-        let backoff = self.retry_config.to_backoff();
-        let hdrs = self.http_downloader.headers();
-
-        // Use backon for retry with exponential backoff and jitter
-        let result = (|| {
-            let client = http_client.clone();
-            let url = url.clone();
-            let segment_path = segment_path.clone();
-            let progress = progress.clone();
-            let rate_limiter = rate_limiter.clone();
-            let hdrs = hdrs.clone();
-
-            async move {
-                // Download segment to file
-                let response = client
-                    .get(&url)
-                    .headers(hdrs)
-                    .timeout(Duration::from_secs(30)) // 30 second timeout per segment
-                    .send()
-                    .await
-                    .map_err(|e| {
-                        if e.is_timeout() {
-                            RdlpError::Network(format!("Segment {idx} timeout"))
-                        } else if e.is_connect() {
-                            RdlpError::Network(format!("Segment {idx} connection failed"))
-                        } else {
-                            RdlpError::Network(format!("Segment {idx} request failed: {e}"))
-                        }
-                    })?;
-
-                if !response.status().is_success() {
-                    return Err(RdlpError::Http {
-                        status: response.status().as_u16(),
-                        reason: format!(
-                            "Segment {idx} returned HTTP {}",
-                            response.status()
-                        ),
-                    });
-                }
-
-                // Stream segment to file with progress tracking
-                let file = File::create(&segment_path).await.map_err(RdlpError::Io)?;
-                let mut writer = BufWriter::with_capacity(buffer_size, file);
-                let mut stream = response.bytes_stream();
-                let mut downloaded = 0u64;
-
-                while let Some(chunk_result) = stream.next().await {
-                    let chunk = chunk_result.map_err(|e| {
-                        RdlpError::Network(format!("Segment {idx} read error: {e}"))
-                    })?;
-
-                    writer.write_all(&chunk).await.map_err(RdlpError::Io)?;
-                    downloaded += chunk.len() as u64;
-
-                    // Update shared progress counter (lock-free atomic)
-                    progress.fetch_add(chunk.len() as u64, Ordering::Relaxed);
-
-                    if let Some(ref limiter) = rate_limiter {
-                        limiter.acquire(chunk.len()).await;
-                    }
-                }
-
-                writer.flush().await.map_err(RdlpError::Io)?;
-
-                Ok((idx, segment_path, downloaded))
-            }
-        })
-        .retry(backoff)
-        .when(is_retryable_error)
-        .notify(|err, dur| {
-            warn!(segment = idx, delay:? = dur; "Segment download failed, retrying: {err}");
-        })
-        .await?;
-
-        Ok(result)
-    }
-
-    /// Download an fMP4 initialization segment, respecting optional byte range.
-    ///
-    /// EXT-X-MAP may specify a `BYTERANGE` when the init data lives inside a
-    /// larger resource.  When present we send an HTTP `Range` header so we only
-    /// fetch the relevant bytes.
-    async fn download_init_segment(
-        &self,
-        init: &InitSegmentInfo,
-        dest: &Path,
-    ) -> Result<()> {
-        let client = self.http_downloader.client().clone();
-        let backoff = self.retry_config.to_backoff();
-        let url = init.url.clone();
-        let byte_range = init.byte_range;
-        let dest = dest.to_path_buf();
-        let hdrs = self.http_downloader.headers();
-
-        (|| {
-            let client = client.clone();
-            let url = url.clone();
-            let dest = dest.clone();
-            let hdrs = hdrs.clone();
-            async move {
-                let mut req = client.get(&url).headers(hdrs).timeout(Duration::from_secs(30));
-
-                // Apply Range header when EXT-X-MAP specifies BYTERANGE
-                if let Some((length, offset)) = byte_range {
-                    let start = offset.unwrap_or(0);
-                    let end = start + length - 1;
-                    req = req.header("Range", format!("bytes={start}-{end}"));
-                    debug!(start, end; "Init segment byte-range request");
-                }
-
-                let response = req.send().await.map_err(|e| {
-                    RdlpError::Network(format!("Init segment request failed: {e}"))
-                })?;
-
-                if !response.status().is_success() {
-                    return Err(RdlpError::Http {
-                        status: response.status().as_u16(),
-                        reason: format!(
-                            "Init segment returned HTTP {}",
-                            response.status()
-                        ),
-                    });
-                }
-
-                let bytes = response.bytes().await.map_err(|e| {
-                    RdlpError::Network(format!("Init segment read error: {e}"))
-                })?;
-
-                tokio::fs::write(&dest, &bytes).await.map_err(RdlpError::Io)?;
-                debug!(bytes = bytes.len(); "Init segment downloaded");
-                Ok(())
-            }
-        })
-        .retry(backoff)
-        .when(is_retryable_error)
-        .notify(|err, dur| {
-            warn!(delay:? = dur; "Init segment download failed, retrying: {err}");
-        })
-        .await
-    }
-
-    /// Parse m3u8 playlist and extract segment URLs
-    ///
-    /// Handles both media playlists (direct segments) and master playlists
-    /// (redirects to best variant). Uses recursive parsing for master playlists.
-    ///
-    /// # Arguments
-    /// * `m3u8_url` - URL of the m3u8 playlist
-    ///
-    /// # Returns
-    /// * `Ok(PlaylistParseResult)` - Segments and optional init segment URL
-    /// * `Err(_)` - Network error, parse error, or empty playlist
-    async fn parse_playlist(&self, m3u8_url: &str) -> Result<PlaylistParseResult> {
-        // Fetch playlist text
-        let playlist_text = self
-            .http_downloader
-            .client()
-            .get(m3u8_url)
-            .headers(self.http_downloader.headers())
-            .send()
-            .await
-            .map_err(|e| RdlpError::Network(format!("Failed to fetch playlist: {e}")))?
-            .text()
-            .await
-            .map_err(|e| RdlpError::Network(format!("Failed to read playlist: {e}")))?;
-
-        // Parse with m3u8-rs
-        let playlist = m3u8_rs::parse_playlist_res(playlist_text.as_bytes()).map_err(|e| {
-            // Show the actual response content when parsing fails (e.g. CDN error pages)
-            if !playlist_text.trim().starts_with("#EXTM3U") {
-                let preview: String = playlist_text.chars().take(200).collect();
-                RdlpError::Extraction(format!(
-                    "Server returned invalid M3U8 (likely expired token or CDN error): {preview}"
-                ))
-            } else {
-                RdlpError::Extraction(format!("M3U8 parse error: {e:?}"))
-            }
-        })?;
-
-        match playlist {
-            m3u8_rs::Playlist::MediaPlaylist(media) => {
-                // Warn about encryption (not yet supported)
-                if media.segments.iter().any(|s| s.key.is_some()) {
-                    warn!(
-                        "HLS stream uses encryption (AES-128/SAMPLE-AES) — decryption not yet supported"
-                    );
-                }
-                // Warn about live streams
-                if !media.end_list {
-                    warn!(
-                        "HLS stream appears to be live (no EXT-X-ENDLIST) — may not download completely"
-                    );
-                }
-
-                // Direct media playlist - extract segments with durations
-                let base_url = url::Url::parse(m3u8_url)
-                    .map_err(|e| RdlpError::Extraction(format!("Invalid base URL: {e}")))?;
-
-                // Build per-segment init info from EXT-X-MAP.
-                // m3u8_rs sets `seg.map` on each segment the tag applies to,
-                // so we just map it directly — handles multiple EXT-X-MAP tags.
-                let segments: Vec<SegmentInfo> = media
-                    .segments
-                    .iter()
-                    .map(|seg| {
-                        let url = base_url
-                            .join(&seg.uri)
-                            .map(|u| u.to_string())
-                            .unwrap_or_else(|_| seg.uri.clone());
-
-                        let init_segment = seg.map.as_ref().map(|map| {
-                            let init_url = base_url
-                                .join(&map.uri)
-                                .map(|u| u.to_string())
-                                .unwrap_or_else(|_| map.uri.clone());
-                            InitSegmentInfo {
-                                url: init_url,
-                                byte_range: map
-                                    .byte_range
-                                    .as_ref()
-                                    .map(|br| (br.length, br.offset)),
-                            }
-                        });
-
-                        SegmentInfo {
-                            url,
-                            duration: seg.duration as f64,
-                            init_segment,
-                        }
-                    })
-                    .collect();
-
-                if segments.is_empty() {
-                    return Err(RdlpError::Extraction("Playlist has no segments".into()));
-                }
-
-                // Log init segment info
-                let unique_inits: HashSet<_> = segments
-                    .iter()
-                    .filter_map(|s| s.init_segment.as_ref())
-                    .collect();
-                if !unique_inits.is_empty() {
-                    debug!(
-                        count = unique_inits.len();
-                        "fMP4 stream detected (EXT-X-MAP init segments)"
-                    );
-                }
-
-                // Security check: limit max segments
-                const MAX_SEGMENTS: usize = 10_000;
-                if segments.len() > MAX_SEGMENTS {
-                    return Err(RdlpError::Extraction(format!(
-                        "Playlist has too many segments: {} (max: {})",
-                        segments.len(),
-                        MAX_SEGMENTS
-                    )));
-                }
-
-                Ok(PlaylistParseResult { segments })
-            }
-            m3u8_rs::Playlist::MasterPlaylist(master) => {
-                // Master playlist - select best quality variant
-                if master.variants.is_empty() {
-                    return Err(RdlpError::Extraction(
-                        "Master playlist has no variants".into(),
-                    ));
-                }
-
-                let variant = master
-                    .variants
-                    .iter()
-                    .filter(|v| !v.is_i_frame)
-                    .max_by_key(|v| v.bandwidth)
-                    .or_else(|| master.variants.iter().max_by_key(|v| v.bandwidth))
-                    .unwrap(); // safe: checked non-empty above
-                let base_url = url::Url::parse(m3u8_url)
-                    .map_err(|e| RdlpError::Extraction(format!("Invalid base URL: {e}")))?;
-
-                let media_playlist_url = base_url
-                    .join(&variant.uri)
-                    .map_err(|e| RdlpError::Extraction(format!("Failed to join URL: {e}")))?
-                    .to_string();
-
-                debug!(
-                    variant:? = variant.uri,
-                    bandwidth = variant.bandwidth;
-                    "Master playlist detected, selecting variant"
-                );
-
-                // Recursively parse media playlist (will detect EXT-X-MAP there)
-                Box::pin(self.parse_playlist(&media_playlist_url)).await
-            }
-        }
-    }
-
-    /// Download segments with resume support
-    ///
-    /// Skips segments that are already completed (tracked in state).
-    /// Updates state after each successful segment download.
-    /// Saves state periodically for crash recovery.
-    ///
-    /// # Arguments
-    /// * `segments` - List of segments with URLs and durations
-    /// * `temp_dir` - Directory to save temporary segment files
-    /// * `base_filename` - Base filename for temporary files
-    /// * `progress_counter` - Shared atomic counter for bytes downloaded
-    /// * `segments_counter` - Shared atomic counter for segments completed
-    /// * `duration_counter` - Shared atomic counter for duration completed (in centiseconds for precision)
-    /// * `state` - Shared download state for resume tracking
-    /// * `output_path` - Final output path (for state file location)
-    ///
-    /// # Returns
-    /// * `Ok(Vec<PathBuf>)` - Paths to ALL segment files (in order, including pre-existing)
-    /// * `Err(_)` - Download error (network, I/O, etc.)
-    #[allow(clippy::too_many_arguments)]
-    #[instrument(skip(self, segments, progress_counter, segments_counter, duration_counter, state), fields(segments = segments.len()))]
-    async fn download_segments_with_resume(
-        &self,
-        segments: Vec<SegmentInfo>,
-        temp_dir: &Path,
-        base_filename: &str,
-        progress_counter: Arc<AtomicU64>,
-        segments_counter: Arc<AtomicU64>,
-        duration_counter: Arc<AtomicU64>,
-        state: Arc<Mutex<HlsDownloadState>>,
-        output_path: &Path,
-    ) -> Result<Vec<PathBuf>> {
-        let total_segments = segments.len();
-
-        // Get already completed segments
-        let completed: HashSet<usize> = state.lock().await.completed_segments.clone();
-        let to_download: Vec<(usize, SegmentInfo)> = segments
-            .iter()
-            .enumerate()
-            .filter(|(idx, _)| !completed.contains(idx))
-            .map(|(idx, seg)| (idx, seg.clone()))
-            .collect();
-
-        // Calculate duration already downloaded from completed segments
-        let completed_duration: f64 = segments
-            .iter()
-            .enumerate()
-            .filter(|(idx, _)| completed.contains(idx))
-            .map(|(_, seg)| seg.duration)
-            .sum();
-        duration_counter.store((completed_duration * 100.0) as u64, Ordering::Relaxed);
-
-        let already_downloaded = completed.len();
-        let remaining = to_download.len();
-        let concurrent = self.concurrent_segments;
-
-        if remaining == 0 {
-            info!(total = total_segments; "All segments already downloaded, skipping to merge");
-        } else {
-            info!(
-                remaining,
-                completed = already_downloaded,
-                concurrent;
-                "Downloading HLS segments"
-            );
-        }
-
-        // Download remaining segments using buffer_unordered with state tracking
-        let downloader = self.clone();
-        let temp_dir_owned = temp_dir.to_path_buf();
-        let base_filename_owned = base_filename.to_string();
-        let output_path_owned = output_path.to_path_buf();
-
-        let max_failures = self.max_segment_failures;
-        let mut stream = stream::iter(to_download.into_iter())
-            .map(|(idx, seg)| {
-                let segment_path = temp_dir_owned.join(format!("{base_filename_owned}.part{idx}"));
-                let downloader = downloader.clone();
-                let progress = progress_counter.clone();
-                let segments = segments_counter.clone();
-                let duration = duration_counter.clone();
-                let state = state.clone();
-                let output_path = output_path_owned.clone();
-                let seg_duration = seg.duration;
-                let seg_url = seg.url;
-
-                async move {
-                    // Check if segment file already exists and is non-empty (single async stat)
-                    if let Ok(meta) = tokio::fs::metadata(&segment_path).await {
-                        if meta.len() > 0 {
-                            debug!(
-                                segment = idx,
-                                bytes = meta.len();
-                                "Segment already exists, skipping"
-                            );
-                            let bytes = meta.len();
-                            // Mark as completed in state
-                            {
-                                let mut state_guard = state.lock().await;
-                                state_guard.mark_completed(idx, bytes);
-                            }
-                            segments.fetch_add(1, Ordering::Relaxed);
-                            progress.fetch_add(bytes, Ordering::Relaxed);
-                            duration.fetch_add((seg_duration * 100.0) as u64, Ordering::Relaxed);
-                            return Ok((idx, segment_path, bytes));
-                        }
-                    }
-
-                    // Download segment with retry logic (now using backon)
-                    let result = downloader
-                        .download_segment_with_retry(
-                            idx,
-                            seg_url,
-                            segment_path.clone(),
-                            progress.clone(),
-                        )
-                        .await;
-
-                    match &result {
-                        Ok((_, _, bytes)) => {
-                            // Update state on success; clone if periodic save needed
-                            let snapshot = {
-                                let mut guard = state.lock().await;
-                                guard.mark_completed(idx, *bytes);
-                                if guard.completed_segments.len() % 50 == 0 {
-                                    Some(guard.clone())
-                                } else {
-                                    None
-                                }
-                            };
-                            // Save outside the lock to avoid holding it across I/O
-                            if let Some(snapshot) = snapshot {
-                                if let Err(e) = snapshot.save(&output_path).await {
-                                    warn!("Failed to save HLS state: {e}");
-                                }
-                            }
-
-                            segments.fetch_add(1, Ordering::Relaxed);
-                            duration.fetch_add((seg_duration * 100.0) as u64, Ordering::Relaxed);
-                        }
-                        Err(e) => {
-                            // Save state on error before propagating
-                            let snapshot = state.lock().await.clone();
-                            if let Err(save_err) = snapshot.save(&output_path).await {
-                                warn!("Failed to save HLS state on error: {save_err}");
-                            }
-                            warn!(segment = idx; "Segment download failed: {e}");
-                        }
-                    }
-
-                    result
-                }
-            })
-            .buffer_unordered(self.concurrent_segments);
-
-        // Collect results, tolerating up to max_segment_failures errors
-        let mut results: Vec<(usize, PathBuf, u64)> = Vec::new();
-        let mut segment_failures = 0usize;
-
-        while let Some(result) = stream.next().await {
-            match result {
-                Ok(item) => results.push(item),
-                Err(e) => {
-                    segment_failures += 1;
-                    warn!(
-                        failures = segment_failures,
-                        max = max_failures;
-                        "Segment failed: {e}"
-                    );
-                    if segment_failures >= max_failures {
-                        let snapshot = state.lock().await.clone();
-                        let _ = snapshot.save(output_path).await;
-                        return Err(RdlpError::Download(format!(
-                            "Too many segment failures ({segment_failures}), aborting"
-                        )));
-                    }
-                }
-            }
-        }
-        if segment_failures > 0 {
-            warn!(failures = segment_failures; "Completed with segment failures");
-        }
-
-        // Save final state (clone under lock, then save outside lock)
-        let snapshot = state.lock().await.clone();
-        if let Err(e) = snapshot.save(output_path).await {
-            warn!("Failed to save final HLS state: {e}");
-        }
-
-        // Build complete list of segment paths (including pre-existing)
-        let mut all_segment_paths: Vec<(usize, PathBuf)> = Vec::with_capacity(total_segments);
-
-        // Add downloaded segments from this run
-        for (idx, path, _) in results {
-            all_segment_paths.push((idx, path));
-        }
-
-        // Add pre-existing segments
-        for idx in completed {
-            let segment_path = temp_dir.join(format!("{base_filename}.part{idx}"));
-            if segment_path.exists() {
-                all_segment_paths.push((idx, segment_path));
-            }
-        }
-
-        // Sort by index for correct merge order
-        all_segment_paths.sort_by_key(|(idx, _)| *idx);
-
-        // Verify we have the expected number of segments (accounting for tolerated failures)
-        let expected_count = total_segments - segment_failures;
-        let actual_count = all_segment_paths.len();
-        if actual_count != expected_count {
-            return Err(RdlpError::Download(format!(
-                "Missing segments: expected {expected_count}, got {actual_count}"
-            )));
-        }
-
-        let segment_paths: Vec<PathBuf> = all_segment_paths
-            .into_iter()
-            .map(|(_, path)| path)
-            .collect();
-
-        info!(total = total_segments; "All segments ready for merge");
-        Ok(segment_paths)
-    }
-
-    /// Merge segment files into final output file.
-    ///
-    /// For fMP4 streams, each segment may reference an init segment (EXT-X-MAP).
-    /// The init segment is written before the first media segment that uses it,
-    /// and re-written whenever the init segment changes (supporting playlists
-    /// with multiple EXT-X-MAP tags).
-    ///
-    /// `segment_init_paths[i]` is the init segment file for `segment_paths[i]`,
-    /// or `None` for plain TS segments.
-    async fn merge_segments(
-        &self,
-        segment_paths: Vec<PathBuf>,
-        output_path: &Path,
-        segment_init_paths: &[Option<PathBuf>],
-    ) -> Result<u64> {
-        let merge_timeout = self.merge_timeout;
-        tokio::time::timeout(merge_timeout, async {
-            let has_init = segment_init_paths.iter().any(|p| p.is_some());
-            info!(
-                segments = segment_paths.len(),
-                fmp4 = has_init;
-                "Merging segments into final file"
-            );
-
-            let final_file = File::create(output_path).await.map_err(RdlpError::Io)?;
-            let mut writer = BufWriter::with_capacity(self.buffer_size, final_file);
-            let mut total_bytes = 0u64;
-
-            // Track the current init segment so we only re-write it on change
-            let mut current_init: Option<&Path> = None;
-
-            for (idx, segment_path) in segment_paths.iter().enumerate() {
-                // Check if this segment needs an init segment (different from current)
-                let seg_init = segment_init_paths
-                    .get(idx)
-                    .and_then(|p| p.as_deref());
-
-                if seg_init != current_init {
-                    if let Some(init_path) = seg_init {
-                        let mut init_file =
-                            File::open(init_path).await.map_err(RdlpError::Io)?;
-                        let bytes = tokio::io::copy(&mut init_file, &mut writer)
-                            .await
-                            .map_err(RdlpError::Io)?;
-                        total_bytes += bytes;
-                        debug!(bytes, segment = idx; "Wrote fMP4 init segment");
-                    }
-                    current_init = seg_init;
-                }
-
-                let mut segment_file =
-                    File::open(segment_path).await.map_err(RdlpError::Io)?;
-                let bytes = tokio::io::copy(&mut segment_file, &mut writer)
-                    .await
-                    .map_err(RdlpError::Io)?;
-                total_bytes += bytes;
-
-                if (idx + 1) % 100 == 0 || idx == segment_paths.len() - 1 {
-                    debug!(
-                        merged = idx + 1,
-                        total = segment_paths.len(),
-                        mb = total_bytes / (1024 * 1024);
-                        "Merge progress"
-                    );
-                }
-            }
-
-            writer.flush().await.map_err(RdlpError::Io)?;
-            info!(mb = total_bytes / (1024 * 1024); "Merge complete");
-
-            Ok(total_bytes)
-        })
-        .await
-        .map_err(|_| {
-            RdlpError::Download(format!(
-                "Merge timed out after {}s",
-                merge_timeout.as_secs()
-            ))
-        })?
-    }
-
-    /// Clean up temporary segment files
-    ///
-    /// Deletes all temporary segment files after successful merge.
-    /// Logs deletion progress for transparency.
-    ///
-    /// # Arguments
-    /// * `segment_paths` - Paths to segment files to delete
-    async fn cleanup_segments(&self, segment_paths: Vec<PathBuf>) {
-        debug!(count = segment_paths.len(); "Cleaning up segment files");
-
-        let mut deleted = 0;
-        for path in segment_paths {
-            if tokio::fs::remove_file(&path).await.is_ok() {
-                deleted += 1;
-            }
-        }
-
-        debug!(deleted; "Segment cleanup complete");
-    }
-}
-
-impl Default for HlsDownloader {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[async_trait]
-impl Downloader for HlsDownloader {
-    fn protocol(&self) -> &str {
-        "hls"
-    }
-
-    fn supports(&self, url: &str) -> bool {
-        url.ends_with(".m3u8") || url.contains("/playlist.m3u8") || url.contains(".m3u8?")
-    }
-
-    async fn get_size(&self, _url: &str) -> Result<Option<u64>> {
-        // Size detection is handled by HlsSizeDetector in extractor layer
-        Ok(None)
-    }
-
-    #[instrument(skip(self, progress), fields(url = %url, path = %path.display()))]
-    async fn download_to_file(
-        &self,
-        url: &str,
-        path: &Path,
-        progress: Option<Box<dyn ProgressCallback>>,
-    ) -> Result<DownloadStats> {
-        let timeout = self.download_timeout;
-        tokio::time::timeout(timeout, async {
-            let start_time = Instant::now();
-
-            // Step 1: Parse playlist
-            let playlist = self.parse_playlist(url).await?;
-            let segments = playlist.segments;
-            let has_init = segments.iter().any(|s| s.init_segment.is_some());
-            let total_segments = segments.len();
-            let total_duration: f64 = segments.iter().map(|s| s.duration).sum();
-            info!(
-                segments = total_segments,
-                duration_secs = total_duration,
-                fmp4 = has_init;
-                "Parsed HLS playlist"
-            );
-
-            // Step 2: Load or create state for resume support
-            let state = match HlsDownloadState::load(path, url, total_segments).await {
-                Some(existing_state) => {
-                    let completed = existing_state.completed_segments.len();
-                    let remaining = total_segments - completed;
-                    info!(
-                        completed,
-                        total = total_segments,
-                        remaining;
-                        "Resuming HLS download"
-                    );
-                    Arc::new(Mutex::new(existing_state))
-                }
-                None => {
-                    info!("Starting fresh HLS download");
-                    Arc::new(Mutex::new(HlsDownloadState::new(
-                        url.to_string(),
-                        total_segments,
-                    )))
-                }
-            };
-
-            // Step 3: Setup progress tracking
-            let downloaded = Arc::new(AtomicU64::new(state.lock().await.total_bytes_downloaded));
-            let segments_completed = Arc::new(AtomicU64::new(
-                state.lock().await.completed_segments.len() as u64,
-            ));
-            // Duration in centiseconds for atomic precision (f64 -> u64)
-            let duration_completed = Arc::new(AtomicU64::new(0));
-            let total_segments_u64 = total_segments as u64;
-
-            // Spawn progress reporter task with duration-based progress
-            let mut progress_guard = spawn_progress_reporter(
-                progress,
-                ProgressMetrics::with_duration(
-                    downloaded.clone(),
-                    segments_completed.clone(),
-                    duration_completed.clone(),
-                ),
-                ProgressReporterConfig::hls(start_time, total_segments_u64, total_duration),
-            );
-
-            // Step 4: Download segments (with resume support)
-            let temp_dir = path.parent().unwrap_or_else(|| Path::new("."));
-            let base_filename = path
-                .file_stem()
-                .and_then(|s| s.to_str())
-                .unwrap_or("download");
-
-            let segment_paths = match self
-                .download_segments_with_resume(
-                    segments.clone(),
-                    temp_dir,
-                    base_filename,
-                    downloaded.clone(),
-                    segments_completed.clone(),
-                    duration_completed.clone(),
-                    state.clone(),
-                    path,
-                )
-                .await
-            {
-                Ok(paths) => paths,
-                Err(e) => {
-                    // Save state on error (so we can resume later)
-                    let snapshot = state.lock().await.clone();
-                    if let Err(save_err) = snapshot.save(path).await {
-                        warn!("Failed to save HLS state: {save_err}");
-                    }
-                    progress_guard.abort();
-                    return Err(e);
-                }
-            };
-
-            // Progress guard will be dropped and abort the task automatically
-            drop(progress_guard);
-
-            // Step 5: Download unique init segments (fMP4 EXT-X-MAP)
-            // Collect unique init segments and download each once.
-            let mut init_file_map: HashMap<InitSegmentInfo, PathBuf> = HashMap::new();
-            {
-                let unique_inits: Vec<InitSegmentInfo> = segments
-                    .iter()
-                    .filter_map(|s| s.init_segment.clone())
-                    .collect::<HashSet<_>>()
-                    .into_iter()
-                    .collect();
-
-                for (i, init) in unique_inits.into_iter().enumerate() {
-                    let init_path = temp_dir.join(format!("{base_filename}.init{i}"));
-                    info!(index = i; "Downloading fMP4 init segment (EXT-X-MAP): {}", init.url);
-                    self.download_init_segment(&init, &init_path).await?;
-                    init_file_map.insert(init, init_path);
-                }
-            }
-
-            // Build per-segment init path mapping for merge
-            let segment_init_paths: Vec<Option<PathBuf>> = segments
-                .iter()
-                .map(|s| {
-                    s.init_segment
-                        .as_ref()
-                        .and_then(|init| init_file_map.get(init).cloned())
-                })
-                .collect();
-
-            // Step 6: Merge segments (re-inserting init segment on change)
-            let total_bytes = self
-                .merge_segments(segment_paths.clone(), path, &segment_init_paths)
-                .await?;
-
-            // Step 7: Cleanup segments, init segments, and state file
-            self.cleanup_segments(segment_paths).await;
-            for init_path in init_file_map.values() {
-                let _ = tokio::fs::remove_file(init_path).await;
-            }
-            if let Err(e) = HlsDownloadState::delete(path).await {
-                warn!("Failed to delete HLS state file: {e}");
-            }
-
-            // Step 8: Return statistics
-            let duration = start_time.elapsed();
-            let stats = DownloadStats::new(total_bytes, duration, 0).with_fragments(total_segments);
-
-            let duration_secs = duration.as_secs_f64();
-            let speed_mbps = (total_bytes as f64 / duration_secs) / (1024.0 * 1024.0);
-            info!(
-                mb = total_bytes / (1024 * 1024),
-                duration_secs:? = duration_secs,
-                speed_mbps:? = speed_mbps;
-                "HLS download complete"
-            );
-
-            Ok(stats)
-        })
-        .await
-        .map_err(|_| {
-            RdlpError::Download(format!("Download timed out after {}s", timeout.as_secs()))
-        })?
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_hls_downloader_creation() {
-        let downloader = HlsDownloader::new();
-        assert_eq!(downloader.protocol(), "hls");
-        assert_eq!(downloader.concurrent_segments, 8);
-    }
-
-    #[test]
-    fn test_hls_downloader_builder() {
-        let downloader = HlsDownloader::new()
-            .with_concurrent_segments(16)
-            .with_buffer_size(4 * 1024 * 1024);
-
-        assert_eq!(downloader.concurrent_segments, 16);
-        assert_eq!(downloader.buffer_size, 4 * 1024 * 1024);
-    }
-
-    #[test]
-    fn test_supports_m3u8_urls() {
-        let downloader = HlsDownloader::new();
-
-        assert!(downloader.supports("https://example.com/video.m3u8"));
-        assert!(downloader.supports("https://example.com/playlist.m3u8"));
-        assert!(downloader.supports("https://example.com/index.m3u8?token=abc"));
-        assert!(!downloader.supports("https://example.com/video.mp4"));
-    }
-
-    #[test]
-    fn test_concurrent_segments_minimum() {
-        let downloader = HlsDownloader::new().with_concurrent_segments(0);
-        // Should be clamped to minimum of 1
-        assert_eq!(downloader.concurrent_segments, 1);
-    }
-}
