+++ crates/rdlp-downloader/src/hls/types.rs
@@ -1,0 +1,54 @@
+//! HLS type definitions for segment and playlist parsing.
+
+use std::path::PathBuf;
+
+/// EXT-X-MAP initialization segment info (fMP4 streams)
+///
+/// Per the HLS spec, `EXT-X-MAP` applies to every segment after it until
+/// the next `EXT-X-MAP` tag. A playlist may therefore contain multiple
+/// init segments (e.g. codec change mid-stream, ad insertion).
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub(crate) struct InitSegmentInfo {
+    /// Fully-resolved URL of the initialization segment
+    pub url: String,
+    /// Optional byte range `(length, offset)` — when the init data is
+    /// packed inside a larger resource (e.g. the same URI as the segments)
+    pub byte_range: Option<(u64, Option<u64>)>,
+}
+
+/// Information about an HLS segment including its duration
+#[derive(Clone, Debug)]
+pub(crate) struct SegmentInfo {
+    /// Segment URL
+    pub url: String,
+    /// Segment duration in seconds (from EXTINF)
+    pub duration: f64,
+    /// The EXT-X-MAP init segment that applies to this segment (if any)
+    pub init_segment: Option<InitSegmentInfo>,
+}
+
+/// Result of parsing an HLS playlist
+#[derive(Clone, Debug)]
+pub(crate) struct PlaylistParseResult {
+    /// Media segments (each carrying its own init segment reference)
+    pub segments: Vec<SegmentInfo>,
+}
+
+/// Segment download result containing index, path, and bytes downloaded
+pub(crate) struct SegmentDownloadResult {
+    pub index: usize,
+    pub path: PathBuf,
+    pub bytes: u64,
+}
+
+impl From<(usize, PathBuf, u64)> for SegmentDownloadResult {
+    fn from((index, path, bytes): (usize, PathBuf, u64)) -> Self {
+        Self { index, path, bytes }
+    }
+}
+
+impl From<SegmentDownloadResult> for (usize, PathBuf, u64) {
+    fn from(result: SegmentDownloadResult) -> Self {
+        (result.index, result.path, result.bytes)
+    }
+}
